requirements.txt
APScheduler==3.9.1.post1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
gvapi==0.8
idna==3.4
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
